Restructuring packages
moving some modules around to try and make the structure a little cleaner
fixed some imports
