+++ snappy/common/cclogging.py
