+++ snappy/common/deserialization_decorator.py
