+++ snappy/os_client.py
@@ -13,7 +13,7 @@ from snappy.tools import randomstring as randstring
 from snappy.tools.md5hash import get_md5_hash
 
 from responses import AccountContainersList, ContainerObjectsList
-from deserialization_decorator import deserialize
+from snappy.common.deserialization_decorator import deserialize
 
 BULK_ARCHIVE_NAME = 'bulk_objects'
