+++ snappy/tests/regression/object/object_range_request_test.py
@@ -1,8 +1,7 @@
 from snappy.fixtures.swift_fixtures import ObjectStorageFixture
 
 from snappy.constants import Constants
-from snappy.decorators import (DataDrivenFixture, data_driven_test)
-from snappy.decorators import DataDrivenFixture
+from snappy.common.decorators import (DataDrivenFixture, data_driven_test)
 from snappy.generators import ObjectDatasetList
